daBossKoopaJrDemo_c: Fix softlock on W6 ship boss
Temporarily disables clown car collision during the boss start demo

+++ source/wiimj2d/d_system/d_a_player_base.cpp
@@ -369,7 +369,7 @@ void daPlBase_c::executeState_DemoControl() {
         onStatus(116);
         if (isNowBgCross(BGC_FOOT)) {
             if (!isState(StateID_Walk) && !isState(StateID_Turn)) {
-                changeState(StateID_Walk, static_cast<long>(AnmBlend_e::DEFAULT));
+                changeState(StateID_Walk, static_cast<int>(AnmBlend_e::DEFAULT));
             }
         } else {
             if (!isState(StateID_Fall)) {
